Move GPIO type definitions to esp-hal-common
