+++ esp-hal-common/src/gpio.rs
@@ -10,6 +10,12 @@ use core::marker::PhantomData;
 
 pub use paste::paste;
 
+#[cfg_attr(feature = "esp32", path = "gpio/esp32.rs")]
+#[cfg_attr(feature = "esp32c3", path = "gpio/esp32c3.rs")]
+#[cfg_attr(feature = "esp32s2", path = "gpio/esp32s2.rs")]
+#[cfg_attr(feature = "esp32s3", path = "gpio/esp32s3.rs")]
+pub mod types;
+
 #[derive(Copy, Clone)]
 pub enum Event {
     RisingEdge  = 1,
