+++ esp-hal-common/src/gpio/esp32s3.rs
@@ -1,0 +1,188 @@
+pub type OutputSignalType = u16;
+pub const OUTPUT_SIGNAL_MAX: u16 = 256;
+pub const INPUT_SIGNAL_MAX: u16 = 189;
+
+/// Peripheral input signals for the GPIO mux
+#[allow(non_camel_case_types)]
+#[derive(PartialEq, Copy, Clone)]
+pub enum InputSignal {
+    SPIQ              = 0,
+    SPID              = 1,
+    SPIHD             = 2,
+    SPIWP             = 3,
+    SPID4             = 7,
+    SPID5             = 8,
+    SPID6             = 9,
+    SPID7             = 10,
+    SPIDQS            = 11,
+    U0RXD             = 12,
+    U0CTS             = 13,
+    U0DSR             = 14,
+    U1RXD             = 15,
+    U1CTS             = 16,
+    U1DSR             = 17,
+    U2RXD             = 18,
+    U2CTS             = 19,
+    U2DSR             = 20,
+    I2S1_MCLK         = 21,
+    I2S0O_BCK         = 22,
+    I2S0_MCLK         = 23,
+    I2S0O_WS          = 24,
+    I2S0I_SD          = 25,
+    I2S0I_BCK         = 26,
+    I2S0I_WS          = 27,
+    I2S1O_BCK         = 28,
+    I2S1O_WS          = 29,
+    I2S1I_SD          = 30,
+    I2S1I_BCK         = 31,
+    I2S1I_WS          = 32,
+    I2S0I_SD1         = 51,
+    I2S0I_SD2         = 52,
+    I2S0I_SD3         = 53,
+    USB_OTG_IDDIG     = 58,
+    USB_OTG_AVALID    = 59,
+    USB_SRP_BVALID    = 60,
+    USB_OTG_VBUSVALID = 61,
+    USB_SRP_SESSEND   = 62,
+    SPI3_CLK          = 66,
+    SPI3_Q            = 67,
+    SPI3_D            = 68,
+    SPI3_HD           = 69,
+    SPI3_WP           = 70,
+    SPI3_CS0          = 71,
+    I2CEXT0_SCL       = 89,
+    I2CEXT0_SDA       = 90,
+    I2CEXT1_SCL       = 91,
+    I2CEXT1_SDA       = 92,
+    FSPICLK           = 101,
+    FSPIQ             = 102,
+    FSPID             = 103,
+    FSPIHD            = 104,
+    FSPIWP            = 105,
+    FSPIIO4           = 106,
+    FSPIIO5           = 107,
+    FSPIIO6           = 108,
+    FSPIIO7           = 109,
+    FSPICS0           = 110,
+    SUBSPIQ           = 120,
+    SUBSPID           = 121,
+    SUBSPIHD          = 122,
+    SUBSPIWP          = 123,
+    SUBSPID4          = 155,
+    SUBSPID5          = 156,
+    SUBSPID6          = 157,
+    SUBSPID7          = 158,
+    SUBSPIDQS         = 159,
+    PWM0_SYNC0        = 160,
+    PWM0_SYNC1        = 161,
+    PWM0_SYNC2        = 162,
+    PWM0_F0           = 163,
+    PWM0_F1           = 164,
+    PWM0_F2           = 165,
+    PWM0_CAP0         = 166,
+    PWM0_CAP1         = 167,
+    PWM0_CAP2         = 168,
+    PWM1_SYNC0        = 169,
+    PWM1_SYNC1        = 170,
+    PWM1_SYNC2        = 171,
+    PWM1_F0           = 172,
+    PWM1_F1           = 173,
+    PWM1_F2           = 174,
+    PWM1_CAP0         = 175,
+    PWM1_CAP1         = 176,
+    PWM1_CAP2         = 177,
+    PCMFSYNC          = 188,
+    PCMCLK            = 189,
+}
+
+/// Peripheral input signals for the GPIO mux
+#[allow(non_camel_case_types)]
+#[derive(PartialEq, Copy, Clone)]
+pub enum OutputSignal {
+    SPIQ            = 0,
+    SPID            = 1,
+    SPIHD           = 2,
+    SPIWP           = 3,
+    SPICLK          = 4,
+    SPICS0          = 5,
+    SPICS1          = 6,
+    SPID4           = 7,
+    SPID5           = 8,
+    SPID6           = 9,
+    SPID7           = 10,
+    SPIDQS          = 11,
+    U0TXD           = 12,
+    U0RTS           = 13,
+    U0DTR           = 14,
+    U1TXD           = 15,
+    U1RTS           = 16,
+    U1DTR           = 17,
+    U2TXD           = 18,
+    U2RTS           = 19,
+    U2DTR           = 20,
+    I2S1_MCLK       = 21,
+    I2S0O_BCK       = 22,
+    I2S0_MCLK       = 23,
+    I2S0O_WS        = 24,
+    I2S0O_SD        = 25,
+    I2S0I_BCK       = 26,
+    I2S0I_WS        = 27,
+    I2S1O_BCK       = 28,
+    I2S1O_WS        = 29,
+    I2S1O_SD        = 30,
+    I2S1I_BCK       = 31,
+    I2S1I_WS        = 32,
+    SPI3_CLK        = 66,
+    SPI3_Q          = 67,
+    SPI3_D          = 68,
+    SPI3_HD         = 69,
+    SPI3_WP         = 70,
+    SPI3_CS0        = 71,
+    SPI3_CS1        = 72,
+    I2CEXT0_SCL     = 89,
+    I2CEXT0_SDA     = 90,
+    I2CEXT1_SCL     = 91,
+    I2CEXT1_SDA     = 92,
+    GPIO_SD0        = 93,
+    GPIO_SD1        = 94,
+    GPIO_SD2        = 95,
+    GPIO_SD3        = 96,
+    GPIO_SD4        = 97,
+    GPIO_SD5        = 98,
+    GPIO_SD6        = 99,
+    GPIO_SD7        = 100,
+    FSPICLK         = 101,
+    FSPIQ           = 102,
+    FSPID           = 103,
+    FSPIHD          = 104,
+    FSPIWP          = 105,
+    FSPIIO4         = 106,
+    FSPIIO5         = 107,
+    FSPIIO6         = 108,
+    FSPIIO7         = 109,
+    FSPICS0         = 110,
+    FSPICS1         = 111,
+    FSPICS2         = 112,
+    FSPICS3         = 113,
+    FSPICS4         = 114,
+    FSPICS5         = 115,
+    SUBSPICLK       = 119,
+    SUBSPIQ         = 120,
+    SUBSPID         = 121,
+    SUBSPIHD        = 122,
+    SUBSPIWP        = 123,
+    SUBSPICS0       = 124,
+    SUBSPICS1       = 125,
+    FSPIDQS         = 126,
+    SPI3_CS2        = 127,
+    I2S0O_SD1       = 128,
+    SUBSPID4        = 155,
+    SUBSPID5        = 156,
+    SUBSPID6        = 157,
+    SUBSPID7        = 158,
+    SUBSPIDQS       = 159,
+    SDIO_TOHOST_INT = 177,
+    PCMFSYNC        = 194,
+    PCMCLK          = 195,
+    GPIO            = 256,
+}
