+++ esp32s2-hal/src/gpio.rs
@@ -1,153 +1,4 @@
-use esp_hal_common::gpio::*;
-
-type OutputSignalType = u16;
-const OUTPUT_SIGNAL_MAX: u16 = 256;
-const INPUT_SIGNAL_MAX: u16 = 204;
-
-/// Peripheral input signals for the GPIO mux
-#[allow(non_camel_case_types)]
-#[derive(PartialEq, Copy, Clone)]
-pub enum InputSignal {
-    SPIQ              = 0,
-    SPID              = 1,
-    SPIHD             = 2,
-    SPIWP             = 3,
-    SPID4             = 7,
-    SPID5             = 8,
-    SPID6             = 9,
-    SPID7             = 10,
-    SPIDQS            = 11,
-    U0RXD             = 14,
-    U0CTS             = 15,
-    U0DSR             = 16,
-    U1RXD             = 17,
-    U1CTS             = 18,
-    U1DSR             = 21,
-    I2S0O_BCK         = 23,
-    I2S0O_WS          = 25,
-    I2S0I_BCK         = 27,
-    I2S0I_WS          = 28,
-    I2CEXT0_SCL       = 29,
-    I2CEXT0_SDA       = 30,
-    USB_OTG_IDDIG     = 64,
-    USB_OTG_AVALID    = 65,
-    USB_SRP_BVALID    = 66,
-    USB_OTG_VBUSVALID = 67,
-    USB_SRP_SESSEND   = 68,
-    SPI3_CLK          = 72,
-    SPI3_Q            = 73,
-    SPI3_D            = 74,
-    SPI3_HD           = 75,
-    SPI3_CS0          = 76,
-    I2CEXT1_SCL       = 95,
-    I2CEXT1_SDA       = 96,
-    FSPICLK           = 108,
-    FSPIQ             = 109,
-    FSPID             = 110,
-    FSPIHD            = 111,
-    FSPIWP            = 112,
-    FSPIIO4           = 113,
-    FSPIIO5           = 114,
-    FSPIIO6           = 115,
-    FSPIIO7           = 116,
-    FSPICS0           = 117,
-    SUBSPIQ           = 127,
-    SUBSPID           = 128,
-    SUBSPIHD          = 129,
-    SUBSPIWP          = 130,
-    SUBSPID4          = 167,
-    SUBSPID5          = 168,
-    SUBSPID6          = 169,
-    SUBSPID7          = 170,
-    SUBSPIDQS         = 171,
-    PCMFSYNC          = 203,
-    PCMCLK            = 204,
-}
-
-/// Peripheral input signals for the GPIO mux
-#[allow(non_camel_case_types)]
-#[derive(PartialEq, Copy, Clone)]
-pub enum OutputSignal {
-    SPIQ            = 0,
-    SPID            = 1,
-    SPIHD           = 2,
-    SPIWP           = 3,
-    SPICLK          = 4,
-    SPICS0          = 5,
-    SPICS1          = 6,
-    SPID4           = 7,
-    SPID5           = 8,
-    SPID6           = 9,
-    SPID7           = 10,
-    SPIDQS          = 11,
-    U0TXD           = 14,
-    U0RTS           = 15,
-    U0DTR           = 16,
-    U1TXD           = 17,
-    U1RTS           = 18,
-    U1DTR           = 21,
-    I2S0O_BCK       = 23,
-    I2S0O_WS        = 25,
-    I2S0I_BCK       = 27,
-    I2S0I_WS        = 28,
-    I2CEXT0_SCL     = 29,
-    I2CEXT0_SDA     = 30,
-    SDIO_TOHOST_INT = 31,
-    SPI3_CLK        = 72,
-    SPI3_Q          = 73,
-    SPI3_D          = 74,
-    SPI3_HD         = 75,
-    SPI3_CS0        = 76,
-    SPI3_CS1        = 77,
-    SPI3_CS2        = 78,
-    I2CEXT1_SCL     = 95,
-    I2CEXT1_SDA     = 96,
-    GPIO_SD0        = 100,
-    GPIO_SD1        = 101,
-    GPIO_SD2        = 102,
-    GPIO_SD3        = 103,
-    GPIO_SD4        = 104,
-    GPIO_SD5        = 105,
-    GPIO_SD6        = 106,
-    GPIO_SD7        = 107,
-    FSPICLK         = 108,
-    FSPIQ           = 109,
-    FSPID           = 110,
-    FSPIHD          = 111,
-    FSPIWP          = 112,
-    FSPIIO4         = 113,
-    FSPIIO5         = 114,
-    FSPIIO6         = 115,
-    FSPIIO7         = 116,
-    FSPICS0         = 117,
-    FSPICS1         = 118,
-    FSPICS2         = 119,
-    FSPICS3         = 120,
-    FSPICS4         = 121,
-    FSPICS5         = 122,
-    SUBSPICLK       = 126,
-    SUBSPIQ         = 127,
-    SUBSPID         = 128,
-    SUBSPIHD        = 129,
-    SUBSPIWP        = 130,
-    SUBSPICS0       = 131,
-    SUBSPICS1       = 132,
-    FSPIDQS         = 133,
-    FSPI_HSYNC      = 134,
-    FSPI_VSYNC      = 135,
-    FSPI_DE         = 136,
-    FSPICD          = 137,
-    SPI3_CD         = 139,
-    SPI3_DQS        = 140,
-    SUBSPID4        = 167,
-    SUBSPID5        = 168,
-    SUBSPID6        = 169,
-    SUBSPID7        = 170,
-    SUBSPIDQS       = 171,
-    PCMFSYNC        = 209,
-    PCMCLK          = 210,
-    GPIO            = 256,
-}
+use esp_hal_common::gpio::{types::*, *};
 
 gpio! {
     Function1,
